+++ common/src/main/java/net/zelythia/autotools/AutoTools.java
@@ -5,7 +5,6 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
@@ -205,26 +204,30 @@ public class AutoTools {
             return;
         }
 
+        sourceSlot = toGuiSlotIndex(sourceSlot);
+
         int destSlot = AutoToolsConfig.get().keepSlot ? inventory.selected : getSuitableHotbarSlot(inventory);
         if (!AutoToolsConfig.get().targetSlots.contains(destSlot + 1)) destSlot = AutoToolsConfig.get().targetSlots.get(0) - 1;
 
-        if(swaps.peek() != sourceSlot) swaps.push(sourceSlot);
+
+        if (swaps.peek() != sourceSlot) swaps.push(sourceSlot);
         if (swaps.peek() != destSlot) swaps.push(destSlot);
 
         swapped = true;
-        if (Screen.hasShiftDown()) {
-            //Simulating a click on the toolSlot and the swappableSlot with the ClickType = SWAP, so it updates with on server
-            client.gameMode.handleInventoryMouseClick(client.player.inventoryMenu.containerId, destSlot + 18, sourceSlot, ClickType.SWAP, client.player);
-            client.gameMode.handleInventoryMouseClick(client.player.inventoryMenu.containerId, destSlot + 27, sourceSlot, ClickType.SWAP, client.player);
-            client.gameMode.handleInventoryMouseClick(client.player.inventoryMenu.containerId, destSlot + 36, sourceSlot, ClickType.SWAP, client.player);
-        } else {
-            client.gameMode.handleInventoryMouseClick(client.player.inventoryMenu.containerId, destSlot + 36, sourceSlot, ClickType.SWAP, client.player);
-        }
+        client.gameMode.handleInventoryMouseClick(client.player.inventoryMenu.containerId, sourceSlot, destSlot, ClickType.SWAP, client.player);
+
 
         inventory.selected = destSlot;
         inventory.setChanged();
 
         if(!AutoToolsConfig.get().switchBack) swaps.clear(); //Easy way to safe some memory because swaps are only needed for switchBack
+    }
+
+    public static int toGuiSlotIndex(int inventoryIndex){
+        if(inventoryIndex <= 8) return inventoryIndex + 36;                             // Hotbar
+        if(inventoryIndex >= 36 && inventoryIndex <= 39) return 44 - inventoryIndex;    // Armor
+        if(inventoryIndex == 40) return 45;                                             // Offhand
+        return inventoryIndex;                                                          // Inventory
     }
 
     /**