+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-- Fix bug that caused crashed on game launch on Forge
-- Fix bug where AutoTools wouldn't work for wood, stone and copper tools while underwater
-- Fix German toggle message
+- Fixed swapping to wrong slots when tool is in offhand (#55)
+- Fixed possible crash during inventory sync (#57)
+- Added better error handling and hints for targetSlots/ignoredSlots config